+++ src/DataSource/DataSourceContextConfig.ts
@@ -1,0 +1,5 @@
+export interface DataSourceContextConfig {
+  id: string;
+  durable: boolean;
+  persisterType: string;
+}