+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './HcTypeOrmModule';
 export * from './Schema/ValueObject';
 export * from './Schema/TypeOrmEntitySchema';
 export * from './Repository';
+export * from './DataSource';