+++ test/helper/Module/Test/Infrastructure/PrivateTestInfraModule.ts
@@ -1,7 +1,10 @@
 import { HcInfraDomainModule } from '@hexancore/core';
 import { Module } from '@nestjs/common';
 
+import * as typeorm from '@test/Module/Test/Infrastructure/Persistance/TypeOrm';
+const t = typeorm;
+
 @Module({
-  imports: [HcInfraDomainModule.forFeature({ moduleInfraFilePath: __dirname })],
+  imports: [HcInfraDomainModule.forFeature({ moduleInfraDir: __dirname })],
 })
 export class PrivateTestInfraModule {}